hopefully made a new rainbow function
