+++ competition/simgui.json
@@ -1,5 +1,13 @@
 {
   "HALProvider": {
+    "Addressable LEDs": {
+      "0": {
+        "columns": 8
+      },
+      "window": {
+        "visible": true
+      }
+    },
     "PWM Outputs": {
       "window": {
         "visible": true
