+++ competition/src/main/java/frc/robot/Led/KwarqsLed.java
@@ -2,13 +2,14 @@ package frc.robot.Led;
 
 public class KwarqsLed {
     private LedController ledController = new LedController(64);
-    private static double m_rainbowFirstPixelHue = 1;
+    
 
     public KwarqsLed() {
         ledController.add("yellow", new Yellow());
         ledController.add("purple", new Purple());
         ledController.add("green", new Green());
         ledController.add("dark", new Dark());
+        ledController.add("rainbow", new Rainbow());
     }
 
     public void disable() {
@@ -28,19 +29,7 @@ public class KwarqsLed {
     } 
 
     public void setRainbow()  {
-         // For every pixel
-    for (var i = 0; i < 64; i++) {
-        // Calculate the hue - hue is easier for rainbows because the color
-        // shape is a circle so only one value needs to precess
-        final var hue = (m_rainbowFirstPixelHue + (i * 180 / 64)) % 180;
-        // Set the value
-        int intValueHue = (int) hue;
-        ledController.ledBuffer.setHSV(i, intValueHue, 255, 128);
-      }
-      // Increase by to make the rainbow "move"
-      m_rainbowFirstPixelHue += 3;
-      // Check bounds
-      m_rainbowFirstPixelHue %= 180;
+       ledController.set("rainbow");
     }
 
     public void run() {
